HAJAE-9 modify dir structure and rename
