+++ src/app/(pages)/about/_components/code-showcase-codebox/code-showcase-codebox.tsx
@@ -1,6 +1,6 @@
 'use client';
 import Image from 'next/image';
-import styles from './codebox.module.scss';
+import styles from './code-showcase-codebox.module.scss';
 import profileImage from '@/public/images/profile_image.png';
 import { RiChatSmile3Fill, RiCloseLine } from '@remixicon/react';
 import Star from '@/ui/star/star';
