+++ src/app/(pages)/about/_components/code-showcase/code-showcase.module.scss
