+++ src/app/(pages)/about/_components/content-body/content-body.module.scss
