+++ src/app/(pages)/about/_components/content-tab/content-tab.module.scss
