+++ src/app/(pages)/about/_components/explorer-content/explorer-content.module.scss
@@ -1,0 +1,33 @@
+.subject {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 5px;
+
+  padding: 8px 5px;
+  border-bottom: solid 1px var(--line);
+
+  color: var(--secondary-white);
+}
+
+.subject-text {
+  margin: 0;
+}
+
+.content {
+  height: 200px;
+  border-bottom: solid 1px var(--line);
+
+  label {
+    color: var(--secondary-white);
+  }
+}
+
+.directories-wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 8px;
+
+  padding: 14px;
+}
