+++ src/app/(pages)/about/_components/explorer-directory/explorer-directory.module.scss
