+++ src/app/(pages)/about/_components/explorer/explorer.tsx
@@ -1,56 +1,12 @@
-'use client';
 import { RiArrowDownSFill, RiMailFill, RiPhoneFill } from '@remixicon/react';
 import styles from './explorer.module.scss';
-import { DIRECTORY_LIST, DIR_COLORS, Directory, MY_EMAIL, MY_PHONE } from '@/types/constants';
-import { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '@/lib/store';
-import { setAboutExplorerDirectoryId } from '@/lib/features/about-explorer-slice';
-import DirectoryComponent from './directory/directory';
+import { MY_EMAIL, MY_PHONE } from '@/types/constants';
+import ExplorerContentComponent from '../explorer-content/explorer-content';
 
 const AboutExplorer: React.FC = () => {
-  const [directories, setDirectories] = useState<Directory[]>([]);
-  const dispatch = useDispatch();
-
-  const explorerState = useSelector((state: RootState) => state.aboutExplorerReducer);
-
-  useEffect(() => {
-    const targetDirectoies = DIRECTORY_LIST.directories.filter(
-      (directory) => directory.directoryType === explorerState.type
-    );
-    if (targetDirectoies) {
-      setDirectories(targetDirectoies);
-      dispatch(setAboutExplorerDirectoryId(targetDirectoies[0].directoryId));
-    }
-  }, [explorerState.type, dispatch]);
-
-  const getSubject = () => {
-    switch (explorerState.type) {
-      case 'terminal':
-        return 'professional-info';
-      case 'user':
-        return 'personal-info';
-      case 'gamepad':
-        return 'hobbies-info';
-    }
-  };
-
   return (
     <div className={styles.explorer}>
-      <div className={styles.subject}>
-        <RiArrowDownSFill color="var(--secondary-white)" size={20} />
-        <label className={styles.subjectText}>{getSubject()}</label>
-      </div>
-      <div>
-        <div className={styles.content}>
-          <div className={styles.directoriesWrapper}>
-            {directories &&
-              directories.map((directory, index) => (
-                <DirectoryComponent key={directory.directoryId} directory={directory} dirColor={DIR_COLORS[index]} />
-              ))}
-          </div>
-        </div>
-      </div>
+      <ExplorerContentComponent />
       <div>
         <div className={styles.contactMe}>
           <RiArrowDownSFill color="var(--secondary-white)" size={20} />
